requirements.txt
gradio
python-dotenv
langchain
langchain-deepseek
langgraph
langchain-anthropic
tavily-python
ipython
streamlit
pytest
selenium
webdriver-manager
